fix: types and methods
