+++ mobile/app/(app)/home/home.tsx
@@ -3,7 +3,7 @@ import { FlatList, TouchableOpacity, View } from "react-native";
 import { useNavigation, useRouter } from "expo-router";
 import { Layout } from "@/components/index";
 import Text from "@/components/text";
-import { loggedIn, selectMasterPassword, useAppDispatch, useAppSelector } from "@/redux";
+import { selectMasterPassword, useAppDispatch, useAppSelector } from "@/redux";
 import { KeyInfo, useGnoNativeContext } from "@gnolang/gnonative";
 import Octicons from '@expo/vector-icons/Octicons';
 import TextInput from "@/components/textinput";
@@ -61,8 +61,6 @@ export default function Page() {
       await gnonative.setPassword(masterPassword);
 
       setLoading(undefined);
-
-      await dispatch(loggedIn({ keyInfo }));
 
       await dispatch(setVaultToEdit({ vault: keyInfo }));
       route.push("/vault/details");
