+++ mobile/redux/features/index.ts
@@ -1,4 +1,2 @@
-export * from "./accountSlice";
-export * from "./profileSlice";
 export * from "./vaultEditSlice";
 export * from "./signinSlice";
