+++ mobile/redux/redux-provider.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Provider } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
-import { accountSlice, profileSlice, vaultSlice, signinSlice } from "./features";
+import { vaultSlice, signinSlice } from "./features";
 import { GnoNativeApi, useGnoNativeContext } from "@gnolang/gnonative";
 import { signUpSlice } from "./features/signupSlice";
 import { useSearch, UseSearchReturnType } from "@/src/hooks/use-search";
@@ -14,19 +14,19 @@ export interface ThunkExtra {
   extra: { gnonative: GnoNativeApi; search: UseSearchReturnType };
 }
 
+export const reducer = {
+  [vaultSlice.reducerPath]: vaultSlice.reducer,
+  [signUpSlice.reducerPath]: signUpSlice.reducer,
+  [signinSlice.reducerPath]: signinSlice.reducer,
+}
+
 const ReduxProvider: React.FC<Props> = ({ children }) => {
   // Exposing GnoNative API to reduxjs/toolkit
   const { gnonative } = useGnoNativeContext();
   const search = useSearch();
 
   const store = configureStore({
-    reducer: {
-      [accountSlice.reducerPath]: accountSlice.reducer,
-      [profileSlice.reducerPath]: profileSlice.reducer,
-      [vaultSlice.reducerPath]: vaultSlice.reducer,
-      [signUpSlice.reducerPath]: signUpSlice.reducer,
-      [signinSlice.reducerPath]: signinSlice.reducer,
-    },
+    reducer,
     middleware: (getDefaultMiddleware) =>
       getDefaultMiddleware({
         serializableCheck: false,
